Implement working with accural orders

+++ models/accural.go
@@ -1,7 +1,7 @@
 package models
 
 type OrderAccrual struct {
-	OrderNumber string `json:"order"`
-	Status      string `json:"status"`
-	Accrual     int64  `json:"accrual"`
+	OrderNumber string  `json:"order"`
+	Status      string  `json:"status"`
+	Accrual     float64 `json:"accrual"`
 }
